refactor(eval): extract the shared readiness-verdict envelope (#4096, epic #4094)
The codepr-enable-readiness and improvement-enforce-readiness evaluators carried
byte-identical ReadinessCriterion / verdict-envelope types. DRY them into one
authoritative readiness-verdict.ts (ReadinessCriterion, ReadinessVerdict).

Behavior-preserving + API-stable: each module re-exports the shared types under its
historical names (CodePrReadinessCriterion/CodePrEnableReadiness, EnforceReadinessReport)
and keeps its own presenceCriterion — the two copies emit deliberately different detail
wording ("no {label}" vs "no named {label}"), a real domain difference left unshared.

Envelope ONLY, by design: a survey of all four corpus->score->verdict sites confirmed
their corpus types, score fns, evidence shapes, and gate configs diverge and stay
per-consumer. Full pipeline extraction deferred to a follow-up once a third corpus-based
gate demonstrates a unified shape. Scope re-cleared by a 3-voter fan-out gate that caught
the presenceCriterion divergence (would have silently changed operator-visible output).

37 consumer tests pass unchanged; tsc clean; producer/consumer gate green.
